working functionality of crud with house and address

+++ src/main/java/com/pma/web/service/HouseService.java
@@ -2,6 +2,7 @@ package com.pma.web.service;
 
 import com.pma.web.model.House;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface HouseService {
@@ -14,7 +15,7 @@ public interface HouseService {
 
     public House updateHouse(long id, House house);
 
-    public House updateCost(long id, float pppm);
+    public House updateCost(long id, BigDecimal pppm);
 
     public House updateRooms(long id, Integer no_of_rooms);
 
@@ -24,7 +25,7 @@ public interface HouseService {
 
     public List<House> getHouseByTenant(long tenantID);
 
-    public List<House> getHouseByCost(float min, float max);
+    public List<House> getHouseByCost(BigDecimal min, BigDecimal max);
 
     public List<House> getHouseByRooms(Integer min, Integer max);
